ImageHandling/image_processing/src/text_to_png_requirements.txt
Pillow
pytest
